Move undo history to Editor
