+++ crates/kas-widgets/src/edit/guard.rs
@@ -87,7 +87,7 @@ pub trait EditGuard: Sized {
     /// This function is called after the text is updated (including by keyboard
     /// input, an undo action or by a message like
     /// [`kas::messages::SetValueText`]). The exceptions are setter methods like
-    /// [`clear`](EditField::clear) and [`set_string`](Editor::set_string).
+    /// [`clear`](Editor::clear) and [`set_string`](Editor::set_string).
     fn edit(edit: &mut EditField<Self>, cx: &mut EventCx, data: &Self::Data) {
         let _ = (edit, cx, data);
     }
